Change direction on parsing/printing the data
Invertible syntax descriptions are fun and all, but the library is not
mature enough to be very usable or provide good error messages. More
over, it seems to make sense to have a two-level parsing (lexing and
parsing?) where the data is first pulled into a very primitive matrix of
fields, and then properly into a list of records.